🔵 first separation of db.Mock into its own module
