+++ cmd/service/commands_file_generation.go
@@ -6,13 +6,14 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/fabmob/playground-standard-covoiturage/cmd/service/db"
 	"github.com/fabmob/playground-standard-covoiturage/cmd/test"
 )
 
 //go:generate bash -c "go test -generate > /dev/null"
 
 var generateTestData bool
-var generatedData = NewMockDB()
+var generatedData = db.NewMockDB()
 var commandsFile = strings.Builder{}
 var serverEnvVar = "SERVER"
 var authEnvVar = "API_TOKEN"
@@ -31,7 +32,7 @@ var hasAlreadyAppended = map[string]bool{}
 
 // appendDataIfGenerated is used to populate the `generatedData` db if the
 // -generate flag is provided
-func appendDataIfGenerated(t *testing.T, mockDB *MockDB) {
+func appendDataIfGenerated(t *testing.T, mockDB *db.Mock) {
 	if _, ok := hasAlreadyAppended[t.Name()]; generateTestData && !ok {
 		hasAlreadyAppended[t.Name()] = true
 		appendData(mockDB, generatedData)
@@ -49,7 +50,7 @@ func appendCmdIfGenerated(t *testing.T, request *http.Request, flags test.Flags,
 	}
 }
 
-func appendData(from *MockDB, to *MockDB) {
+func appendData(from *db.Mock, to *db.Mock) {
 	to.DriverJourneys = append(
 		to.GetDriverJourneys(),
 		from.GetDriverJourneys()...,
