+++ cmd/service/db/db.go
@@ -1,4 +1,4 @@
-package service
+package db
 
 import (
 	"bytes"
@@ -13,8 +13,8 @@ import (
 	"github.com/google/uuid"
 )
 
-// MockDB stores the data of the server in memory
-type MockDB struct {
+// Mock stores the data of the server in memory
+type Mock struct {
 	DriverJourneys    []api.DriverJourney
 	PassengerJourneys []api.PassengerJourney
 	Bookings          BookingsByID
@@ -25,8 +25,8 @@ type MockDB struct {
 type BookingsByID map[uuid.UUID]*api.Booking
 
 // NewMockDB initiates a MockDB with no data
-func NewMockDB() *MockDB {
-	m := MockDB{}
+func NewMockDB() *Mock {
+	m := Mock{}
 	m.DriverJourneys = []api.DriverJourney{}
 	m.PassengerJourneys = []api.PassengerJourney{}
 	m.Bookings = BookingsByID{}
@@ -34,7 +34,7 @@ func NewMockDB() *MockDB {
 	return &m
 }
 
-func (m *MockDB) GetDriverJourneys() []api.DriverJourney {
+func (m *Mock) GetDriverJourneys() []api.DriverJourney {
 	if m.DriverJourneys == nil {
 		m.DriverJourneys = []api.DriverJourney{}
 	}
@@ -42,7 +42,7 @@ func (m *MockDB) GetDriverJourneys() []api.DriverJourney {
 	return m.DriverJourneys
 }
 
-func (m *MockDB) GetPassengerJourneys() []api.PassengerJourney {
+func (m *Mock) GetPassengerJourneys() []api.PassengerJourney {
 	if m.PassengerJourneys == nil {
 		m.PassengerJourneys = []api.PassengerJourney{}
 	}
@@ -50,7 +50,7 @@ func (m *MockDB) GetPassengerJourneys() []api.PassengerJourney {
 	return m.PassengerJourneys
 }
 
-func (m *MockDB) GetBookings() BookingsByID {
+func (m *Mock) GetBookings() BookingsByID {
 	if m.Bookings == nil {
 		m.Bookings = BookingsByID{}
 	}
@@ -58,7 +58,7 @@ func (m *MockDB) GetBookings() BookingsByID {
 	return m.Bookings
 }
 
-func (m *MockDB) GetUsers() []api.User {
+func (m *Mock) GetUsers() []api.User {
 	if m.Users == nil {
 		m.Users = []api.User{}
 	}
@@ -66,7 +66,7 @@ func (m *MockDB) GetUsers() []api.User {
 	return m.Users
 }
 
-func (m *MockDB) GetBooking(bookingID uuid.UUID) (*api.Booking, error) {
+func (m *Mock) GetBooking(bookingID uuid.UUID) (*api.Booking, error) {
 	bookings := m.GetBookings()
 
 	booking, ok := bookings[bookingID]
@@ -79,7 +79,7 @@ func (m *MockDB) GetBooking(bookingID uuid.UUID) (*api.Booking, error) {
 
 // AddBooking adds a new booking to the data. Returns an error if a booking
 // with same ID already exists
-func (m *MockDB) AddBooking(booking api.Booking) error {
+func (m *Mock) AddBooking(booking api.Booking) error {
 	bookings := m.GetBookings()
 
 	if _, bookingExists := bookings[booking.Id]; bookingExists {
@@ -91,39 +91,10 @@ func (m *MockDB) AddBooking(booking api.Booking) error {
 	return nil
 }
 
-// UpdateBookingStatus updates the status of a booking. Status can only be
-// updated for a higher ranked status. If this is not the case, or if the
-// booking is not found, returns an error
-func (m *MockDB) UpdateBookingStatus(bookingID uuid.UUID, newStatus api.BookingStatus) error {
-	booking, err := m.GetBooking(bookingID)
-	if err != nil {
-		return err
-	}
-
-	statusAfter, err := statusIsAfter(newStatus, booking.Status)
-	if err != nil {
-		return err
-	}
-
-	if !statusAfter {
-		return StatusAlreadySetErr{}
-	}
-
-	booking.Status = newStatus
-
-	return nil
-}
-
 type MissingBookingErr struct{}
 
 func (err MissingBookingErr) Error() string {
 	return "missing_booking"
-}
-
-type StatusAlreadySetErr struct{}
-
-func (err StatusAlreadySetErr) Error() string {
-	return "status_already_set"
 }
 
 //////////////////////////////////////////////////////////
@@ -138,7 +109,7 @@ type mockDBDataInterface struct {
 	Messages          []api.PostMessagesJSONBody `json:"messages"`
 }
 
-func toOutputData(m *MockDB) mockDBDataInterface {
+func ToOutputData(m *Mock) mockDBDataInterface {
 	outputData := mockDBDataInterface{}
 
 	outputData.DriverJourneys = m.DriverJourneys
@@ -154,7 +125,7 @@ func toOutputData(m *MockDB) mockDBDataInterface {
 	return outputData
 }
 
-func fromInputData(inputData mockDBDataInterface) *MockDB {
+func fromInputData(inputData mockDBDataInterface) *Mock {
 	var m = NewMockDB()
 
 	m.DriverJourneys = inputData.DriverJourneys
@@ -172,13 +143,13 @@ func fromInputData(inputData mockDBDataInterface) *MockDB {
 }
 
 // NewMockDBWithDefaultData initiates a MockDB with default data
-func NewMockDBWithDefaultData() *MockDB {
+func NewMockDBWithDefaultData() *Mock {
 	return MustReadDefaultData()
 }
 
 // NewMockDBWithData reads journey data from io.Reader with json data.
 // It does not validate data against the standard.
-func NewMockDBWithData(r io.Reader) (*MockDB, error) {
+func NewMockDBWithData(r io.Reader) (*Mock, error) {
 	var data mockDBDataInterface
 
 	bytes, readErr := io.ReadAll(r)
@@ -197,7 +168,7 @@ func NewMockDBWithData(r io.Reader) (*MockDB, error) {
 var DefaultData []byte
 
 // MustReadDefaultData reads default data, and panics if any error occurs
-func MustReadDefaultData() *MockDB {
+func MustReadDefaultData() *Mock {
 	mockDB, err := NewMockDBWithData(bytes.NewReader(DefaultData))
 	if err != nil {
 		panic(err)
